finish shapes

+++ Shapes/Maths.h
@@ -63,8 +63,8 @@ struct Color
 bool operator== (const Color& a, const Color& b) { return (a.r == b.r && a.g == b.g && a.b == b.b); }
 //bool operator== (Color& a, Color& b) { if (a.r == b.r && a.g == b.g && a.b == b.b) return true; return false; }
 bool operator< (const Color a, const Color b) { 
-	float l1 = (a.r * 256 + a.g) * 256 + a.b;
-	float l2 = (a.r * 256 + a.g) * 256 + a.b;
+	int l1 = (a.r * 256 + a.g) * 256 + a.b;
+	int l2 = (a.r * 256 + a.g) * 256 + a.b;
 	return (l1 < l2) ;
 }  // needed for use in map
 std::ostream& operator << (std::ostream& out, const Color& c)
